services/ai-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
prophet==1.1.6
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
anthropic==0.43.0
pydantic==2.10.4
python-dotenv==1.0.1
